picking up feature wip
I got the player to pick up the specific sushi when in front of any stall, but after picking it up, only that sushi stall is able to function. The others freeze because the scripts that they all share assume that the "heldSushi" variable contains something when it only applies to that particular sushi stall. Need to move this over to the playerController script