fixed graphics and customer collision
Graphics bug fixed: As the customer was moving, the speech bubble and the sushi orders would not render very smoothly. To fix this, I replaced Translate with modifying the position.x directly.

Customer now detects whether the correct sushi is fed, and checkmarks indicate when that particular sushi request is fulfilled. Once the sushi order is complete (all sushi have checkmarks on them), then the customer walk back